// backend/src/roles/enums/resource.enum.ts
export enum Resource {
  // --- Nhóm Bán hàng ---
  products = 'products',
  categories = 'categories',
  orders = 'orders',
  users = 'users', // Customer và Staff

  // --- Nhóm Marketing ---
  blogs = 'blogs',
  coupons = 'coupons',

  // --- Nhóm Kho vận ---
  inventory = 'inventory',
  suppliers = 'suppliers',

  // --- Nhóm CSKH ---
  contacts = 'contacts',

  // --- Nhóm Hệ thống ---
  statistics = 'statistics',
  roles = 'roles',
  audit_logs = 'audit_logs',
  departments = 'departments',
}